Add files via upload
Initial commit. 
Works on WS2012R2
Warning a bug is still there on this version : the third script is meant to be launched from $env:USERPROFILE\ but since the second script adds you in the Domain, the "C:\Users\Administrator" turns out to be "C:\Users\Administrator.DOMAIN"
Therefore, the next version will create, in the first script, both an installation folder no more linked to the User but rather on a specific folder.